Moving withAuth to server folder
