+++ src/pages/api/auth/profile.ts
@@ -3,7 +3,7 @@ import jwt from 'jsonwebtoken';
 import { getUser } from '../../../data/users';
 import { decrypt } from '../../../auth/server/encrypt';
 import { requireEnvVar } from '../../../errors/envcheck'; 
-import { withAuth } from '../../../middleware/withAuth';
+import { withAuth } from '../../../auth/server/withAuth';
 import { USER_AUTH_TOKEN_NAME } from '../../../../constants';
 
 const SECRET_KEY_JWT = requireEnvVar('SECRET_KEY_JWT');
