+++ src/pages/api/comics/new.ts
@@ -1,5 +1,5 @@
 import { NextApiHandler } from 'next';
-import { withAuth } from '../../middleware/withAuth';
+import { withAuth } from '../../../auth/server/withAuth';
 
 const handler: NextApiHandler = async (req, res) => {
   return res.status(200).send("ok");
